Added more to cart form
Changed the "Checkout" link to say "View Cart". Added a total cost instance variable to be used. Added a dropdown for tax numbers and added a JS function to get the values stored in the select dropdown. The function then rounds to the nearest thousandth place and outputs the total of games price plus taxes.